도움됐어요 2차

+++ src/main/java/com/kh/lawservice101/knowledgein/model/vo/InBoardVo.java
@@ -18,5 +18,5 @@ public class InBoardVo {
     private CategoryVo categoryVo;
     private ClientVo clientVo;
     private String writerDate;
-    private int inBoardHelpCount;
+    private Long inBoardHelpCount;
 }
